+++ docs/changelog.md
@@ -6,3 +6,4 @@
 * Migrating to PSR-4
 * Base class architect
 * View and Routing mapping
+* Online PDF HTML editor